More cleanup

+++ lib/harmonicSolver.cpp
@@ -137,7 +137,7 @@ double handleEdge(const HarmCacheMap &cache, const HarmCacheCIt &it, double tKey
 
 // Solve the simple harmonic for the given time value
 Vec3 harmonicSolver(double time,
-        unsigned int waves, unsigned int length, double amp, double decay,
+        double waves, double length, double amp, double decay,
 		double term,
         const Vec3 &ampAxis, bool matchVelocity, const HarmCacheMap &cache,
         bool ignoreInitialAccel
